move Conf into a common module
